Floor heating devices

+++ ClimateController/Devices/TemperatureSensor.cs
@@ -2,5 +2,5 @@ namespace ClimateController.Devices;
 
 internal abstract class TemperatureSensor
 {
-    public abstract Task<float> GetTemperatureAsync(CancellationToken cancellationToken = default);
+    public abstract Task<decimal> GetTemperatureAsync(CancellationToken cancellationToken = default);
 }
